Mangle DNS queries to CIDR resource resolvers in `StubResolver`
